OneVarFunction.java
This class contains a method that is a test used in OneVariable.java to check whether a random variable classifies as a success. This test can be changed depending on the simulation being done eg for a standard coin flip this class would test whether the random variable is greater than 0.5.